backend/pyproject.toml
[project]
name = "meridian-chatbot"
version = "0.1.0"
description = "Meridian chatbot FastAPI backend"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.136.1",
    "mangum>=0.21.0",
    "openai>=2.33.0",
    "openai-agents>=0.14.8",
    "psycopg[binary]>=3.2",
    "pydantic-settings>=2.14.0",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.46.0",
]
